Allow for newer versions of gcc to compile MeTA.
Tested with GCC 4.8. This is mostly a "hack" in the sense that it falls
back on some things that are in theory less performant for the sake of
getting it to work. It's not a hack in the sense that these features are
detected with CMake and eventually we can eliminate them when GCC catches
up with the standard...

+++ test/vocabulary_map_test.h
@@ -80,7 +80,7 @@ void assert_correctness(uint16_t size = 20)
             // checks that the inverse map contains the correct position of
             // the terms in the lowest level
             if (idx < 14)
-                ASSERT(inverse[idx] == file.tellg());
+                ASSERT(inverse[idx] == static_cast<uint64_t>(file.tellg()));
 
             std::string term;
             uint64_t num;
